requirements.txt
flask>=2.2.0
gunicorn>=20.1.0
pandas>=1.3.0
python-binance>=1.0.16
python-dotenv>=0.21.0
ta>=0.10.2
